moke/requirements.txt
aenum==3.1.15
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
Flask==3.0.2
frozenlist==1.4.1
future==1.0.0
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
line-bot-sdk==3.9.0
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.16.0
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
